requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
pymysql==1.1.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
bcrypt==4.3.0
pytest==8.4.1
httpx==0.28.1
cryptography==45.0.5
alembic==1.16.4